% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compengine.R
\name{AC_9}
\alias{AC_9}
\title{Autocorrelation at lag 9}
\usage{
AC_9(y)
}
\arguments{
\item{y}{the input time series}
}
\value{
autocorrelation at lag 9
}
\description{
Autocorrelation at lag 9
}
\references{
B.D. Fulcher and N.S. Jones. hctsa: A computational framework for automated time-series phenotyping using massive feature extraction. Cell Systems 5, 527 (2017).

B.D. Fulcher, M.A. Little, N.S. Jones Highly comparative time-series analysis: the empirical structure of time series and their methods. J. Roy. Soc. Interface 10, 83 (2013).
}
\author{
Yangzhuoran Yang
}
