% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{acf1}
\alias{acf1}
\title{Autocorrelation function at lag 1}
\usage{
acf1(x)
}
\arguments{
\item{x}{a univariate time series}
}
\value{
A numeric value.
}
\description{
Computes the first order autocorrelation a time series.
}
