% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yanfei.R
\name{arch_test}
\alias{arch_test}
\title{ARCH LM Test}
\usage{
arch_test(x, lags = 12, demean = TRUE)
}
\arguments{
\item{x}{a univariate time series}

\item{lags}{Number of lags to use in the test}

\item{demean}{Should data have mean removed before test applied?}
}
\value{
A numeric value.
}
\description{
Computes the p-value for the Lagrange Multiplier (LM) test of Engle (1982) for
autoregressive conditional heteroscedasticity (ARCH).
}
