% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.list.mts.R
\name{as.list.mts}
\alias{as.list.mts}
\title{Convert mts object to list of time series}
\usage{
\method{as.list}{mts}(x, ...)
}
\arguments{
\item{x}{multivariate time series of class mts.}

\item{...}{other arguments are ignored.}
}
\value{
A list of ts objects.
}
\description{
An mts object contains a multivariate time series in a matrix, with time on rows.
This is converted into a list of univariate time series.
}
\author{
Rob J Hyndman
}
