% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compengine.R
\name{BF_iszscored}
\alias{BF_iszscored}
\title{Crude check for whether a data vector is (eps-close to being) z-scored.}
\usage{
BF_iszscored(x)
}
\arguments{
\item{x}{the input time series (or any vector)}
}
\value{
a logical with the verdict.
#' @references B.D. Fulcher and N.S. Jones. hctsa: A computational framework for automated time-series phenotyping using massive feature extraction. Cell Systems 5, 527 (2017).
}
\description{
Used for displaying warning messages for functions that require z-scored inputs.
}
\references{
B.D. Fulcher, M.A. Little, N.S. Jones Highly comparative time-series analysis: the empirical structure of time series and their methods. J. Roy. Soc. Interface 10, 83 (2013).
}
\author{
Yangzhuoran Yang
}
