% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compengine.R
\name{dist_features}
\alias{dist_features}
\title{The distribution feature set from software package \code{hctsa}}
\usage{
dist_features(x)
}
\arguments{
\item{x}{the input time series}
}
\value{
a vector with distribution features
}
\description{
Calculate the features that grouped as distribution set,
which have been used in CompEngine database, using method introduced in package \code{hctsa}.
}
\details{
Features in this set are \code{histogram_mode_10}
and \code{outlierinclude_mdrmd}.
}
\references{
B.D. Fulcher and N.S. Jones. hctsa: A computational framework for automated time-series phenotyping using massive feature extraction. Cell Systems 5, 527 (2017).

B.D. Fulcher, M.A. Little, N.S. Jones Highly comparative time-series analysis: the empirical structure of time series and their methods. J. Roy. Soc. Interface 10, 83 (2013).
}
\seealso{
\code{\link{histogram_mode}}

\code{\link{outlierinclude_mdrmd}}
}
\author{
Yangzhuoran Yang
}
