% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplacian.R
\name{embedding}
\alias{embedding}
\title{Finding embedding based on distances}
\usage{
embedding(distances, k = 2, method = c("Laplacian", "Lrw", "Lsym", "Lsym2",
  "MDS", "MDSiso", "monoMDS", "DPM", "Rtsne"), h = median(distances), ...)
}
\arguments{
\item{distances}{an object of class "dist" (essential the lower triangle of distances matrix). 
If it is a rectangular matrix, distances are computed.}

\item{k}{embedding dimension}

\item{h}{bandwidth for computing the similarity matrix. Only used for Laplacian methods,
apart from LaplacianMDS where hs is set to a large h.}

\item{...}{any other arguments are ignored.}
}
\description{
Various methods for embedding a distance matrix into a low-dimensional matrix
}
\examples{
mylist <- list(sunspot.year, WWWusage, AirPassengers, USAccDeaths)
fmat <- tsfeatures(mylist)
z <- embedding(dist(fmat))
plot(z)
}
