% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{firstmin_ac_cpp}
\alias{firstmin_ac_cpp}
\title{First local minimum of the ACF}
\usage{
firstmin_ac_cpp(y)
}
\arguments{
\item{y}{the input time series}
}
\description{
First local minimum of the ACF
}
