% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{firstzero_ac_cpp}
\alias{firstzero_ac_cpp}
\title{First zero crossing of the autocorrelation function}
\usage{
firstzero_ac_cpp(y)
}
\arguments{
\item{y}{the input time series}
}
\description{
First zero crossing of the autocorrelation function
}
