% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{flat_spots}
\alias{flat_spots}
\title{Longest flat spot}
\usage{
flat_spots(x)
}
\arguments{
\item{x}{a univariate time series}
}
\value{
A numeric value.
}
\description{
"Flat spots” are computed by dividing the sample space of a time series into ten equal-sized intervals, and computing the maximum run length within any single interval.
}
\author{
Earo Wang and Rob J Hyndman
}
