% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compengine.R
\name{hctsa_autocorr}
\alias{hctsa_autocorr}
\title{The autocorrelation feature set from software package \code{hctsa}}
\usage{
hctsa_autocorr(x)
}
\arguments{
\item{y}{the input time series}
}
\value{
a vector with autocorrelation features
}
\description{
Calculate the features that grouped as autocorrelation set, 
which have been used in CompEngine database, using method introduced in package \code{kctsa}.
}
\details{
Features in this set are \code{CO_Embed2_Basic_tau_incircle_1}, 
\code{CO_Embed2_Basic_tau_incircle_2}, 
\code{AC_9}, 
\code{CO_FirstMin_ac}, 
\code{CO_trev_1_num}, 
\code{SB_MotifTwo_mean_hhh}, 
and \code{PH_Walker_prop_01_sw_propcross}.
}
\references{
B.D. Fulcher and N.S. Jones. hctsa: A computational framework for automated time-series phenotyping using massive feature extraction. Cell Systems 5, 527 (2017).

B.D. Fulcher, M.A. Little, N.S. Jones Highly comparative time-series analysis: the empirical structure of time series and their methods. J. Roy. Soc. Interface 10, 83 (2013).
}
\seealso{
\code{\link{CO_Embed2_Basic_tau_incircle}}

\code{\link{AC_9}}

\code{\link{CO_FirstMin_ac}}

\code{\link{CO_trev_1_num}}

\code{\link{SB_MotifTwo_mean_hhh}}

\code{\link{PH_Walker_prop_01_sw_propcross}}
}
\author{
Yangzhuoran Yang
}
