% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compengine.R
\name{hctsa_pred}
\alias{hctsa_pred}
\title{The prediction feature set from software package \code{hctsa}}
\usage{
hctsa_pred(x)
}
\arguments{
\item{y}{the input time series}
}
\value{
a vector with autocorrelation features
}
\description{
Calculate the features that grouped as prediction set, 
which have been used in CompEngine database, using method introduced in package \code{kctsa}.
}
\details{
Features in this set are \code{FC_LocalSimple_mean1_taures}, 
\code{FC_LocalSimple_lfit_taures}, 
and \code{EN_SampEn_5_03_sampen1}.
}
\references{
B.D. Fulcher and N.S. Jones. hctsa: A computational framework for automated time-series phenotyping using massive feature extraction. Cell Systems 5, 527 (2017).

B.D. Fulcher, M.A. Little, N.S. Jones Highly comparative time-series analysis: the empirical structure of time series and their methods. J. Roy. Soc. Interface 10, 83 (2013).
}
\seealso{
\code{\link{FC_LocalSimple_taures}}

\code{\link{EN_SampEn_5_03_sampen1}}
}
\author{
Yangzhuoran Yang
}
