% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compengine.R
\name{hctsa_scal}
\alias{hctsa_scal}
\title{The scaling feature set from software package \code{hctsa}}
\usage{
hctsa_scal(x)
}
\arguments{
\item{y}{the input time series}
}
\value{
a vector with autocorrelation features
}
\description{
Calculate the features that grouped as scaling set, 
which have been used in CompEngine database, using method introduced in package \code{kctsa}.
}
\details{
Feature in this set is \code{SC_FluctAnal_2_rsrangefit_50_1_logi_prop_r1}.
}
\references{
B.D. Fulcher and N.S. Jones. hctsa: A computational framework for automated time-series phenotyping using massive feature extraction. Cell Systems 5, 527 (2017).

B.D. Fulcher, M.A. Little, N.S. Jones Highly comparative time-series analysis: the empirical structure of time series and their methods. J. Roy. Soc. Interface 10, 83 (2013).
}
\seealso{
\code{\link{SC_FluctAnal_2_rsrangefit_50_1_logi_prop_r1}}
}
\author{
Yangzhuoran Yang
}
