% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compengine.R
\name{hctsa_station}
\alias{hctsa_station}
\title{The stationarity feature set from software package \code{hctsa}}
\usage{
hctsa_station(x)
}
\arguments{
\item{y}{the input time series}
}
\value{
a vector with autocorrelation features
}
\description{
Calculate the features that grouped as stationarity set, 
which have been used in CompEngine database, using method introduced in package \code{kctsa}.
}
\details{
Features in this set are \code{SY_StdNthDer_1}, 
\code{SY_SpreadRandomLocal_50_100_meantaul}, 
and \code{SY_SpreadRandomLocal_ac2_100_meantaul}.
}
\references{
B.D. Fulcher and N.S. Jones. hctsa: A computational framework for automated time-series phenotyping using massive feature extraction. Cell Systems 5, 527 (2017).

B.D. Fulcher, M.A. Little, N.S. Jones Highly comparative time-series analysis: the empirical structure of time series and their methods. J. Roy. Soc. Interface 10, 83 (2013).
}
\seealso{
\code{\link{SY_StdNthDer_1}}

\code{\link{SY_SpreadRandomLocal_100_meantaul}}
}
\author{
Yangzhuoran Yang
}
