% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yanfei.R
\name{heterogeneity}
\alias{heterogeneity}
\title{Heterogeneity coefficients}
\usage{
heterogeneity(x)
}
\arguments{
\item{x}{a univariate time series}
}
\value{
A vector of numeric values.
}
\description{
Computes various measures of heterogeneity of a time series.
}
