% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compengine.R
\name{localsimple_taures}
\alias{localsimple_taures}
\title{The first zero crossing of the autocorrelation function of the residuals from Simple local time-series forecasting from software package \code{hctsa}}
\usage{
localsimple_taures(y, forecastMeth = c("mean", "lfit"), trainLength = NULL)
}
\arguments{
\item{y}{the input time series}

\item{forecastMeth}{the forecasting method, default to \code{mean}.
\code{mean}: local mean prediction using the past trainLength time-series values.
\code{lfit}: local linear prediction using the past trainLength time-series values.}

\item{trainLength}{the number of time-series values to use to forecast the next value.
Default to 1 when using method \code{mean} and 3 when using method \code{lfit}.}
}
\value{
The first zero crossing of the autocorrelation function of the residuals
}
\description{
Simple predictors using the past trainLength values of the time series to
predict its next value.
}
