% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{localsimple_taures_cpp}
\alias{localsimple_taures_cpp}
\title{Local simple forecasting residual tau}
\usage{
localsimple_taures_cpp(y, forecastMeth = "mean", trainLength = -1L)
}
\arguments{
\item{y}{the input time series}

\item{forecastMeth}{the forecasting method, defaults to \code{mean}.}

\item{trainLength}{the training length. Defaults to \code{-1}}
}
\description{
Local simple forecasting residual tau
}
