% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{lumpiness}
\alias{lumpiness}
\alias{stability}
\title{Autocorrelation function at lag 1}
\usage{
lumpiness(x, width = ifelse(frequency(x) > 1, frequency(x), 10))

stability(x, width = ifelse(frequency(x) > 1, frequency(x), 10))
}
\arguments{
\item{x}{a univariate time series}
}
\value{
A numeric value.
}
\description{
Computes the first order autocorrelation a time series.
}
\author{
Rob J Hyndman
}
