% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{lyapunov}
\alias{lyapunov}
\title{Lyapunov coefficent}
\usage{
lyapunov(x)
}
\arguments{
\item{x}{a univariate time series}
}
\value{
A numeric value.
}
\description{
Computes the Lyapunov coefficient indicating the level of nonlinearity of
of a time series.
}
