% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleseasonal.R
\name{multiseasonal}
\alias{multiseasonal}
\title{Multiple seasonality features}
\usage{
multiseasonal(x, transform = FALSE, lambda = NULL, ...)
}
\arguments{
\item{x}{a univariate time series of class \code{msts} or \code{ts}}

\item{transform}{A logical variable indicating if a Box-Cox transform should be applied
before the STL decomposition. If \code{lambda} is not NULL, then \code{transform} is set to TRUE.}

\item{lambda}{The value of the Box-Cox transformation parameter.
If \code{lambda=NULL} and \code{transform=TRUE}, then lambda is automatically selected
using \code{\link[forecast]{BoxCox.lambda}}.}

\item{...}{Other arguments are passed to \code{\link[forecast]{BoxCox.lambda}}.}
}
\value{
A vector of numeric values.
}
\description{
Computes various measures of multiple seasonality of a time series. The number
of seasonal periods, and the length of the seasonal periods are return. Also,
the strength of seasonality corresponding to each period is estimated using
a linear model involving Fourier terms corresponding to the seasonal periods.
This is analogous to the \code{stl_features} approach, but using a linear 
decomposition rather than an STL decomposition.
}
