% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampen_first_cpp}
\alias{sampen_first_cpp}
\title{C++ replacement for sampen_first}
\usage{
sampen_first_cpp(y)
}
\arguments{
\item{y}{the input time series}
}
\description{
C++ replacement for sampen_first
}
