% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampenc_cpp}
\alias{sampenc_cpp}
\title{Sample entropy counting}
\usage{
sampenc_cpp(y, M = 6L, r = 0.3)
}
\arguments{
\item{y}{the input time series}

\item{M}{embedding dimension}

\item{r}{threshold}
}
\description{
Sample entropy counting
}
