% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compengine.R
\name{SB_MotifTwo_mean_hhh}
\alias{SB_MotifTwo_mean_hhh}
\title{Local motifs in a binary symbolization of the time series}
\usage{
SB_MotifTwo_mean_hhh(y)
}
\arguments{
\item{y}{the input time series}
}
\value{
Entropie of words in the binary alphabet of length 3.
}
\description{
Coarse-graining is performed. Time-series values above its mean are given 1, 
and those below the mean are 0.
}
\examples{
SB_MotifTwo_mean_hhh(WWWusage)
}
\references{
B.D. Fulcher and N.S. Jones. hctsa: A computational framework for automated time-series phenotyping using massive feature extraction. Cell Systems 5, 527 (2017).

B.D. Fulcher, M.A. Little, N.S. Jones Highly comparative time-series analysis: the empirical structure of time series and their methods. J. Roy. Soc. Interface 10, 83 (2013).
}
\author{
Yangzhuoran Yang
}
