% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spreadrandomlocal_meantaul_cpp}
\alias{spreadrandomlocal_meantaul_cpp}
\title{Bootstrap stationarity measure}
\usage{
spreadrandomlocal_meantaul_cpp(y, l, starts)
}
\arguments{
\item{y}{the input time series}

\item{l}{the length of local time-series segments to analyse as a positive integer. Can also be a specified character string: "ac2": twice the first zero-crossing of the autocorrelation function}

\item{starts}{start indices}
}
\description{
Bootstrap stationarity measure
}
