% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compengine.R
\name{std1st_der}
\alias{std1st_der}
\title{Standard deviation of the first derivative of the time series from software package \code{hctsa}}
\usage{
std1st_der(y)
}
\arguments{
\item{y}{the input time series. Missing values will be removed.}
}
\value{
Standard deviation of the first derivative of the time series.
}
\description{
Modified from \code{SY_StdNthDer} in \code{kctsa}. Based on an idea by Vladimir Vassilevsky.
}
\references{
cf. http://www.mathworks.de/matlabcentral/newsreader/view_thread/136539

B.D. Fulcher and N.S. Jones. hctsa: A computational framework for automated time-series phenotyping using massive feature extraction. Cell Systems 5, 527 (2017).

B.D. Fulcher, M.A. Little, N.S. Jones Highly comparative time-series analysis: the empirical structure of time series and their methods. J. Roy. Soc. Interface 10, 83 (2013).
}
\author{
Yangzhuoran Yang
}
