% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{stl_features}
\alias{stl_features}
\title{Strength of trend and seasonality of a time series}
\usage{
stl_features(x, s.window = 15, robust = FALSE, transform = FALSE,
  lambda = NULL, ...)
}
\arguments{
\item{x}{a univariate time series}

\item{robust}{A logical variable indicating if robust STL should be applied.}

\item{transform}{A logical variable indicating if a Box-Cox transform should be applied
before the STL decomposition. If \code{lambda} is not NULL, then \code{transform} is set to TRUE.}

\item{lambda}{The value of the Box-Cox transformation parameter.
If \code{lambda=NULL} and \code{transform=TRUE}, then lambda is automatically selected
using \code{\link[forecast]{BoxCox.lambda}}.}

\item{...}{Other arguments are passed to \code{\link[forecast]{BoxCox.lambda}}.}
}
\value{
A numeric value.
}
\description{
Computes the strength of trend and seasonality of a time series using an STL
decomposition (or for non-seasonal time series, a penalized regression spline smoother).
}
