% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsfeatures-package.R
\docType{package}
\name{tsfeatures-package}
\alias{tsfeatures-package}
\alias{_PACKAGE}
\title{tsfeatures: Time Series Feature Extraction}
\description{
Methods for extracting various features from time series data. The features provided are those from Hyndman, Wang and Laptev (2013) \doi{10.1109/ICDMW.2015.104}, Kang, Hyndman and Smith-Miles (2017) \doi{10.1016/j.ijforecast.2016.09.004} and from Fulcher, Little and Jones (2013) \doi{10.1098/rsif.2013.0048}. Features include spectral entropy, autocorrelations, measures of the strength of seasonality and trend, and so on. Users can also define their own feature functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.robjhyndman.com/tsfeatures/}
  \item \url{https://github.com/robjhyndman/tsfeatures}
  \item Report bugs at \url{https://github.com/robjhyndman/tsfeatures/issues}
}

}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID})

Authors:
\itemize{
  \item Yanfei Kang (\href{https://orcid.org/0000-0001-8769-6650}{ORCID})
  \item Pablo Montero-Manso \email{p.montero.manso@udc.es}
  \item Mitchell O'Hara-Wild (\href{https://orcid.org/0000-0001-6729-7695}{ORCID})
  \item Thiyanga Talagala (\href{https://orcid.org/0000-0002-0656-9789}{ORCID})
  \item Earo Wang (\href{https://orcid.org/0000-0001-6448-5260}{ORCID})
  \item Yangzhuoran Yang \email{Fin.Yang@monash.edu}
}

Other contributors:
\itemize{
  \item Souhaib Ben Taieb [contributor]
  \item Cao Hanqing [contributor]
  \item D K Lake [contributor]
  \item Nikolay Laptev [contributor]
  \item J R Moorman [contributor]
  \item Bohan Zhang [contributor]
}

}
\keyword{internal}
