% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{unitroot_kpss}
\alias{unitroot_kpss}
\alias{unitroot_pp}
\title{Unit Root Test Statistics}
\usage{
unitroot_kpss(x, ...)

unitroot_pp(x, ...)
}
\arguments{
\item{x}{a univariate time series.}

\item{...}{Other arguments are passed to the \code{\link[urca]{ur.kpss}} or
\code{\link[urca]{ur.kpss}} functions.}
}
\value{
A numeric value
}
\description{
\code{unitroot_kpss} computes the statistic for the Kwiatkowski et al. unit root test
using the default settings for the \code{\link[urca]{ur.kpss}} function.
\code{unitroot_pp} computes the statistic for the Phillips-Perron unit root test
using the default settings for the \code{\link[urca]{ur.pp}} function.
}
\author{
Pablo Montero-Manso
}
