% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{walker_propcross_cpp}
\alias{walker_propcross_cpp}
\title{Hypothetical walker crossing fraction}
\usage{
walker_propcross_cpp(y, p = 0.1)
}
\arguments{
\item{y}{the input time series}

\item{p}{the particle gap}
}
\description{
Hypothetical walker crossing fraction
}
