% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo.R
\name{yahoo_data}
\alias{yahoo_data}
\title{Yahoo server metrics}
\usage{
yahoo_data(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{download.file}

Downloads and returns aggregated and anonymized datasets from Yahoo representing server metrics of Yahoo services.}
}
\value{
A matrix of time series with 1437 rows of hourly data, and 1748 columns representing different servers.
}
\description{
Yahoo server metrics
}
\examples{
yahoo <- yahoo_data()
plot(yahoo[,1:10])
plot(yahoo[,1:44], plot.type='single', col=1:44)

}
\references{
Hyndman, R.J., Wang, E., Laptev, N. (2015) Large-scale unusual time series detection.
In: \emph{Proceedings of the IEEE International Conference on Data Mining}. Atlantic City, NJ, USA. 14–17 November 2015.
\url{https://robjhyndman.com/publications/icdm2015/}
}
\author{
Rob Hyndman, Earo Wang, Nikolay Laptev, Mitchell O'Hara-Wild
}
