% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thiyanga.R
\name{zero_proportion}
\alias{zero_proportion}
\title{Proportion of zeros}
\usage{
zero_proportion(x, tol = 1e-08)
}
\arguments{
\item{x}{a univariate time series}

\item{tol}{tolerance level. Absolute values below this are considered zeros.}
}
\value{
A numeric value.
}
\description{
Computes proportion of zeros in a time series
}
\author{
Thiyanga Talagala
}
