% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsfeatures.R
\name{as.list.mts}
\alias{as.list.mts}
\title{Convert mts object to list of time series}
\usage{
\method{as.list}{mts}(x, ...)
}
\arguments{
\item{x}{multivariate time series of class mts.}

\item{...}{other arguments are ignored.}
}
\description{
Convert mts object to list of time series
}
\author{
Rob J Hyndman
}
