% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thiyanga.R
\name{holt_parameters}
\alias{holt_parameters}
\alias{hw_parameters}
\title{Parameter estimates of Holt's linear trend method}
\usage{
holt_parameters(x)

hw_parameters(x)
}
\arguments{
\item{x}{a univariate time series}
}
\value{
\code{holt_parameters} produces a vector of 2 values: alpha, beta.

\code{hw_parameters} produces a vector of 3 values: alpha, beta and gamma.
}
\description{
Estimate the smoothing parameter for the level-alpha and
the smoothing parameter for the trend-beta.
\code{hw_parameters} considers additive seasonal trend: ets(A,A,A) model.
}
\author{
Thiyanga Talagala, Pablo Montero-Manso
}
