% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yanfei.R
\name{nonlinearity}
\alias{nonlinearity}
\title{Nonlinearity coefficent}
\usage{
nonlinearity(x)
}
\arguments{
\item{x}{a univariate time series}
}
\value{
A numeric value.
}
\description{
Computes the p-value for Terasvirta's nonlinearity test of a time series.
}
\author{
Yanfei Kang
}
