% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yanfei.R
\name{nonlinearity}
\alias{nonlinearity}
\title{Nonlinearity coefficient}
\usage{
nonlinearity(x)
}
\arguments{
\item{x}{a univariate time series}
}
\value{
A numeric value.
}
\description{
Computes a nonlinearity statistic based on Teräsvirta's nonlinearity test of a time series.
The statistic is \eqn{10X^2/T}{10X^2/T} where \eqn{X^2}{X^2} is the Chi-squared statistic from
Teräsvirta's test, and T is the length of the time series. This takes large values
when the series is nonlinear, and values around 0 when the series is linear.
}
\examples{
nonlinearity(lynx)
}
\author{
Yanfei Kang and Rob J Hyndman
}
