% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{stl_features}
\alias{stl_features}
\title{Strength of trend and seasonality of a time series}
\usage{
stl_features(x)
}
\arguments{
\item{x}{a univariate time series}
}
\value{
A numeric value.
}
\description{
Computes the strength of trend and seasonality of a time series using an STL
decomposition (or for non-seasonal time series, a penalized regression spline smoother).
}
