% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleseasonal.R
\name{stl_features}
\alias{stl_features}
\title{Strength of trend and seasonality of a time series}
\usage{
stl_features(x, ...)
}
\arguments{
\item{x}{a univariate time series.}

\item{...}{Other arguments are passed to \code{\link[forecast]{mstl}}.}
}
\value{
A vector of numeric values.
}
\description{
Computes various measures of trend and seasonality of a time series based on
an STL decomposition. The number of seasonal periods, and the length of the
seasonal periods are returned. Also, the strength of seasonality corresponding
to each period is estimated. The \code{\link[forecast]{mstl}} function is used
to do the decomposition.
}
\author{
Rob J Hyndman
}
