% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{unitroot_kpss}
\alias{unitroot_kpss}
\alias{unitroot_pp}
\title{Unit Root Test Statistics}
\usage{
unitroot_kpss(x)

unitroot_pp(x)
}
\arguments{
\item{x}{a univariate time series.}
}
\value{
A numeric value
}
\description{
\code{unitroot_kpss} computes the statistic for the Kwiatkowski et al. unit root test with linear trend and lag 1.
\code{unitroot_pp} computes the statistic for the `'Z-alpha'' version of Phillips & Perron unit root test with constant trend and lag 1.
}
\author{
Pablo Montero-Manso
}
